extract more utilities for working with config dicts
